feat(bitfield): add `mycelium-bitfield` crate
apparently people are interested in using the `mycelium_util::bits`
module in other projects, so it would be nice to eventually release it
to crates.io as its own standalone crate. this branch factors out that
code from `mycelium-util` into a new `mycelium-bitfield` crate that
doesn't include any other mycelium code that people may not need to
depend on.

this isn't quite ready to release yet, as the code still needs
documentation, API polish, and more tests, but it's a start!
