+++ bitfield/Cargo.toml
@@ -1,11 +1,16 @@
 [package]
 name = "mycelium-bitfield"
 version = "0.1.0"
-authors = ["Eliza Weisman <eliza@elizas.website>"]
-edition = "2018"
+edition = "2021"
+authors = [
+    "Eliza Weisman <eliza@elizas.website>",
+]
+license = "MIT"
 rust-version = "1.61.0"
 
+[dev-dependencies]
+proptest = "1"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
